added logging to files for easier motion planner debugging

+++ motion_planner.h
@@ -116,7 +116,7 @@ public:
     bool full();
     bool empty();
     bool data_ready();
-    uint32_t interpolate(int max_items, uint16_t *buf);
+    uint32_t interpolate(uint32_t max_items, uint16_t *buf);
     void recalculate(void);
     int size();
     void pop();
@@ -129,10 +129,10 @@ private:
     bool _pro_vv(float vs, float ve, std::vector<float> &ad); //return the A/D profile (t1, t2, t3) in vector ad
     float _pro_ad(float vs, float ve, float dm, std::vector<float> &ad);
     float _v0(float t, std::vector<float> &ad);
-    int calculate_delta(float cartesian[NUM_AXIS], uint32_t delta[NUM_AXIS]);
+    int calculate_delta(float cartesian[NUM_AXIS], long int delta[NUM_AXIS]);
     float _v_bar(float t, std::vector<float> &ad, float vs, float vm);
     void zero(void);
-    uint32_t last_pos[NUM_AXIS];
+    long int last_pos[NUM_AXIS];
     boost::circular_buffer<step_t> cb;
 };
